copilot/pyproject.toml
[project]
name = "copilot"
version = "0.1.0"
description = "A LangChain-powered smart home copilot with WoT registry integration via MCP"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3",
    "langchain-openai>=1.1",
    "langchain-mcp-adapters>=0.2",
    "langgraph>=0.4",
    "langgraph-checkpoint-sqlite>=2.0",
    "copilotkit>=0.1.31",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "aiosqlite>=0.20",
    "langfuse>=2.0",
    "pydantic-settings>=2.7",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["copilot"]

[tool.ruff]
target-version = "py312"
line-length = 100
